chore: move prove post inspect to provider
