+++ src/extension/content-script/index.tsx
@@ -1,6 +1,5 @@
 import './injections/Welcome' // ? Inject welcome
 import './injections/Posts' // ? Inject all posts
-import './injections/ProfilePage' // ? Inject to ProfilePage
 import './tasks' // ? AutomatedTabTask Run tasks when invoked by background page
 
 import * as HoloflowsKit from '@holoflows/kit'
